build: add Makefile for building

+++ human.cpp
@@ -30,21 +30,21 @@ void Human::saveGame()
     file << "Board: " << endl;
     file << "Size: " << board.getSize().first << "x" << board.getSize().second << endl;
     file << "Squares: " << endl;
-    for (int i = 0; i < board.getSquares().size(); i++)
+    for (size_t i = 0; i < board.getSquares().size(); i++)
     {
-      for (int j = 0; j < board.getSquares()[i].size(); j++)
+      for (size_t j = 0; j < board.getSquares()[i].size(); j++)
       {
         file << board.getSquares()[i][j] << " ";
       }
       file << endl;
     }
     file << "Ships: " << endl;
-    for (int i = 0; i < ships.size(); i++)
+    for (size_t i = 0; i < ships.size(); i++)
     {
       file << "Ship " << i + 1 << ": " << endl;
       file << "Size: " << ships[i].getSize() << endl;
       file << "Coordinates: ";
-      for (int j = 0; j < ships[i].getCoordinates().size(); j++)
+      for (size_t j = 0; j < ships[i].getCoordinates().size(); j++)
       {
         file << "(" << ships[i].getCoordinates()[j].first << ", " << ships[i].getCoordinates()[j].second << ") ";
       }
